Address review comments: extract _csv_fixtures.py, fix StopIteration, fail on missing data, fix README grammar
